project/pyproject.toml
[project]
name = "data-explorer"
version = "0.1.0"
description = "AI Data Explorer for CSV and Databases"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai==1.0.9",
    "streamlit>=1.52.2",
    "pandas>=2.2.0",
    "sqlalchemy>=2.0.0",
    "openai>=1.108.2",
    "minsearch>=0.0.5",
    "plotly>=5.18.0",
    "matplotlib>=3.8.0",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "tqdm>=4.66.0",
    "pydantic>=2.0.0",
]

